requirements.txt
scipy==0.13.3
numpy==1.8.2
matplotlib==1.3.1
Pillow==2.3.0
pyserial==2.6